feat(files/cosntants): modify constants to be able to create more in order way
