+++ packages/files/src/utils/constants/index.js
@@ -1,0 +1,9 @@
+const tipoDocumentoObj = require('./tipo-documento');
+const tipoEntidadObj = require('./tipo-entidad');
+const uploadDir = require('./upload-dir-path');
+
+module.exports = {
+  tipoDocumentoObj,
+  tipoEntidadObj,
+  uploadDir,
+};
