+++ packages/files/src/utils/constants/tipo-documento.js
@@ -1,42 +1,3 @@
-const tipoEntidadObj = [
-  {
-    id: 1,
-    name: 'INSTITUCION',
-  },
-  {
-    id: 2,
-    name: 'RATIFICACION',
-  },
-  {
-    id: 3,
-    name: 'PERSONA',
-  },
-  {
-    id: 4,
-    name: 'REPRESENTANTE',
-  },
-  {
-    id: 5,
-    name: 'PROGRAMA',
-  },
-  {
-    id: 6,
-    name: 'TRAYECTORIA',
-  },
-  {
-    id: 7,
-    name: 'PLANTEL',
-  },
-  {
-    id: 8,
-    name: 'SOLICITUD',
-  },
-  {
-    id: 9,
-    name: 'EVALUACION',
-  },
-];
-
 const tipoDocumentoObj = [
   {
     id: 1,
@@ -192,4 +153,4 @@ const tipoDocumentoObj = [
   },
 ];
 
-module.exports = { tipoEntidadObj, tipoDocumentoObj };
+module.exports = tipoDocumentoObj;
