+++ packages/files/src/utils/constants/upload-dir-path.js
@@ -1,0 +1,6 @@
+const { constants } = require('@siiges-services/shared');
+const { join } = require('path');
+
+const uploadDir = join(constants.rootDir, 'public');
+
+module.exports = uploadDir;
